refactor(atelier_sfc): split compile_script/props into cohesive modules
