--- crates/vize_atelier_sfc/src/compile_script/props.rs
@@ -1,2144 +1,0 @@
-//! Props and emit type extraction utilities.
-//!
-//! This module handles extracting prop types from TypeScript type definitions
-//! and processing withDefaults defaults.
-
-use oxc_allocator::Allocator;
-use oxc_ast::ast::{
-    Argument, Expression, ObjectPropertyKind, PropertyKey, Statement, TSLiteral,
-    TSMappedTypeModifierOperator, TSSignature, TSType, TSTypeLiteral, TSTypeName,
-    TSTypeOperatorOperator,
-};
-use oxc_parser::Parser;
-use oxc_span::{GetSpan, SourceType};
-use vize_carton::FxHashMap;
-use vize_carton::{String, ToCompactString};
-
-use crate::script::ScriptCompileContext;
-use crate::types::SfcError;
-
-/// Prop type information
-#[derive(Debug, Clone)]
-pub struct PropTypeInfo {
-    /// JavaScript type constructor name (String, Number, Boolean, Array, Object, Function)
-    pub js_type: String,
-    /// Original TypeScript type (for PropType<T> usage)
-    pub ts_type: Option<String>,
-    /// Whether the prop is optional
-    pub optional: bool,
-    /// Whether the prop accepts null at runtime
-    pub nullable: bool,
-}
-
-/// Strip TypeScript comments from source while preserving string literals.
-fn strip_ts_comments(input: &str) -> String {
-    let mut result = String::with_capacity(input.len());
-    let bytes = input.as_bytes();
-    let mut i = 0;
-    let mut in_string = false;
-    let mut string_char = b'"';
-
-    while i < bytes.len() {
-        if in_string {
-            if bytes[i] == string_char && (i == 0 || bytes[i - 1] != b'\\') {
-                in_string = false;
-            }
-            result.push(bytes[i] as char);
-            i += 1;
-            continue;
-        }
-
-        match bytes[i] {
-            b'\'' | b'"' | b'`' => {
-                in_string = true;
-                string_char = bytes[i];
-                result.push(bytes[i] as char);
-                i += 1;
-            }
-            b'/' if i + 1 < bytes.len() && bytes[i + 1] == b'/' => {
-                // Line comment: skip until newline
-                while i < bytes.len() && bytes[i] != b'\n' {
-                    i += 1;
-                }
-            }
-            b'/' if i + 1 < bytes.len() && bytes[i + 1] == b'*' => {
-                // Block comment: skip until */
-                i += 2;
-                while i + 1 < bytes.len() && !(bytes[i] == b'*' && bytes[i + 1] == b'/') {
-                    i += 1;
-                }
-                if i + 1 < bytes.len() {
-                    i += 2; // skip */
-                }
-            }
-            _ => {
-                result.push(bytes[i] as char);
-                i += 1;
-            }
-        }
-    }
-    result
-}
-
-/// Join multi-line type definitions where continuation lines start with `|` or `&`.
-/// For example:
-/// ```text
-/// type?:
-///     | 'input'
-///     | 'text';
-/// ```
-/// becomes: `type?: | 'input' | 'text';`
-fn join_union_continuation_lines(input: &str) -> String {
-    let lines: Vec<&str> = input.lines().collect();
-    let mut result = String::with_capacity(input.len());
-    for (i, line) in lines.iter().enumerate() {
-        let trimmed = line.trim();
-        if trimmed.starts_with('|') || trimmed.starts_with('&') {
-            // Join to previous line with a space
-            result.push(' ');
-            result.push_str(trimmed);
-        } else {
-            if i > 0 {
-                result.push('\n');
-            }
-            result.push_str(line);
-        }
-    }
-    result
-}
-
-/// Extract prop types from TypeScript type definition.
-/// Returns a Vec to preserve definition order (important for matching Vue's output).
-pub fn extract_prop_types_from_type(type_args: &str) -> Vec<(String, PropTypeInfo)> {
-    extract_prop_types_from_type_with_context(type_args, None, None)
-}
-
-pub(crate) fn extract_prop_types_from_type_with_context(
-    type_args: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-) -> Vec<(String, PropTypeInfo)> {
-    if let Some(props) = extract_prop_types_from_ast(type_args, interfaces, type_aliases) {
-        return props;
-    }
-
-    extract_prop_types_from_type_text(type_args)
-}
-
-fn extract_prop_types_from_type_text(type_args: &str) -> Vec<(String, PropTypeInfo)> {
-    let mut props = Vec::new();
-
-    // Strip comments before parsing
-    let stripped = strip_ts_comments(type_args);
-    // Join multi-line union/intersection types (lines starting with | or &)
-    let joined = join_union_continuation_lines(&stripped);
-    let content = joined.trim();
-    let content = if content.starts_with('{') && content.ends_with('}') {
-        &content[1..content.len() - 1]
-    } else {
-        content
-    };
-
-    // Split by commas/semicolons/newlines in a single character pass. Keeping
-    // `prev` avoids building a `Vec<char>` just to look behind for `=>`, which
-    // used to allocate on every type-literal prop extraction.
-    let mut depth: i32 = 0;
-    let mut current = String::default();
-    let mut prev = '\0';
-
-    for c in content.chars() {
-        match c {
-            '{' | '<' | '(' | '[' => {
-                depth += 1;
-                current.push(c);
-            }
-            '}' | ')' | ']' => {
-                if depth > 0 {
-                    depth -= 1;
-                }
-                current.push(c);
-            }
-            '>' => {
-                // Don't count `>` as closing angle bracket when preceded by `=` (arrow function `=>`)
-                if prev == '=' {
-                    current.push(c);
-                } else {
-                    if depth > 0 {
-                        depth -= 1;
-                    }
-                    current.push(c);
-                }
-            }
-            ',' | ';' if depth <= 0 => {
-                extract_prop_type_info(&current, &mut props);
-                current.clear();
-                depth = 0;
-            }
-            '\n' if depth <= 0 => {
-                // Don't split on newline if the current segment ends with ':' (type on next line)
-                let trimmed_current = current.trim();
-                if !trimmed_current.is_empty() && !trimmed_current.ends_with(':') {
-                    extract_prop_type_info(&current, &mut props);
-                    current.clear();
-                    depth = 0;
-                }
-                // If ends with ':', keep accumulating (type continues on next line)
-            }
-            _ => current.push(c),
-        }
-        prev = c;
-    }
-    extract_prop_type_info(&current, &mut props);
-
-    props
-}
-
-const TYPE_ALIAS_PREFIX: &str = "type __VizeProps = ";
-
-fn extract_prop_types_from_ast(
-    type_args: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-) -> Option<Vec<(String, PropTypeInfo)>> {
-    let source = wrap_type_alias_source(type_args);
-    let allocator = Allocator::default();
-    let parsed = Parser::new(&allocator, &source, SourceType::ts()).parse();
-    if parsed.panicked || !parsed.errors.is_empty() {
-        return None;
-    }
-
-    let Some(Statement::TSTypeAliasDeclaration(alias)) = parsed.program.body.first() else {
-        return None;
-    };
-
-    let mut props = Vec::new();
-    let mut seen = Vec::new();
-    if collect_props_from_ts_type(
-        &alias.type_annotation,
-        &source,
-        interfaces,
-        type_aliases,
-        &mut seen,
-        &mut props,
-    ) {
-        Some(props)
-    } else {
-        None
-    }
-}
-
-fn wrap_type_alias_source(type_args: &str) -> String {
-    let trimmed = type_args.trim();
-    let mut source = String::with_capacity(TYPE_ALIAS_PREFIX.len() + trimmed.len() + 1);
-    source.push_str(TYPE_ALIAS_PREFIX);
-    source.push_str(trimmed);
-    source.push(';');
-    source
-}
-
-fn collect_props_from_ts_type(
-    ts_type: &TSType<'_>,
-    source: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    seen: &mut Vec<String>,
-    props: &mut Vec<(String, PropTypeInfo)>,
-) -> bool {
-    match ts_type {
-        TSType::TSTypeLiteral(type_lit) => {
-            collect_props_from_ts_type_literal(type_lit, source, interfaces, type_aliases, props);
-            true
-        }
-        TSType::TSMappedType(mapped) => {
-            collect_props_from_mapped_type(mapped, source, interfaces, type_aliases, seen, props);
-            true
-        }
-        TSType::TSIntersectionType(intersection) => {
-            let mut handled = false;
-            for ty in intersection.types.iter() {
-                handled |=
-                    collect_props_from_ts_type(ty, source, interfaces, type_aliases, seen, props);
-            }
-            handled
-        }
-        TSType::TSUnionType(union) => {
-            let mut handled = false;
-            for ty in union.types.iter() {
-                handled |=
-                    collect_props_from_ts_type(ty, source, interfaces, type_aliases, seen, props);
-            }
-            handled
-        }
-        TSType::TSConditionalType(conditional) => {
-            let true_handled = collect_props_from_ts_type(
-                &conditional.true_type,
-                source,
-                interfaces,
-                type_aliases,
-                seen,
-                props,
-            );
-            let false_handled = collect_props_from_ts_type(
-                &conditional.false_type,
-                source,
-                interfaces,
-                type_aliases,
-                seen,
-                props,
-            );
-            true_handled || false_handled
-        }
-        TSType::TSParenthesizedType(paren) => collect_props_from_ts_type(
-            &paren.type_annotation,
-            source,
-            interfaces,
-            type_aliases,
-            seen,
-            props,
-        ),
-        TSType::TSTypeReference(type_ref) => {
-            let Some(name) = simple_type_name(&type_ref.type_name) else {
-                return false;
-            };
-            let Some(resolved) = resolve_type_reference_text(name, interfaces, type_aliases, seen)
-            else {
-                return false;
-            };
-            let resolved_source = wrap_type_alias_source(&resolved);
-            let allocator = Allocator::default();
-            let parsed = Parser::new(&allocator, &resolved_source, SourceType::ts()).parse();
-            if parsed.panicked || !parsed.errors.is_empty() {
-                finish_resolved_type_reference(name, seen);
-                return false;
-            }
-            let Some(Statement::TSTypeAliasDeclaration(alias)) = parsed.program.body.first() else {
-                finish_resolved_type_reference(name, seen);
-                return false;
-            };
-            let handled = collect_props_from_ts_type(
-                &alias.type_annotation,
-                &resolved_source,
-                interfaces,
-                type_aliases,
-                seen,
-                props,
-            );
-            finish_resolved_type_reference(name, seen);
-            handled
-        }
-        _ => false,
-    }
-}
-
-fn collect_props_from_ts_type_literal(
-    type_lit: &TSTypeLiteral<'_>,
-    source: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    props: &mut Vec<(String, PropTypeInfo)>,
-) {
-    for member in type_lit.members.iter() {
-        match member {
-            TSSignature::TSPropertySignature(prop) => {
-                let Some(name) = property_key_name(&prop.key) else {
-                    continue;
-                };
-                let (ts_type, js_type, nullable) = if let Some(type_ann) = &prop.type_annotation {
-                    let ts_type = source_for_span(source, type_ann.type_annotation.span())
-                        .unwrap_or("unknown")
-                        .trim()
-                        .to_compact_string();
-                    let js_type = ts_type_to_js_type_from_ast(
-                        &type_ann.type_annotation,
-                        source,
-                        interfaces,
-                        type_aliases,
-                    );
-                    let nullable = type_includes_top_level_null_from_ast(&type_ann.type_annotation);
-                    (Some(ts_type), js_type, nullable)
-                } else {
-                    (None, "null".to_compact_string(), false)
-                };
-                push_prop_type_info(
-                    props,
-                    name,
-                    PropTypeInfo {
-                        js_type,
-                        ts_type,
-                        optional: prop.optional,
-                        nullable,
-                    },
-                );
-            }
-            TSSignature::TSMethodSignature(method) => {
-                let Some(name) = property_key_name(&method.key) else {
-                    continue;
-                };
-                push_prop_type_info(
-                    props,
-                    name,
-                    PropTypeInfo {
-                        js_type: "Function".to_compact_string(),
-                        ts_type: None,
-                        optional: method.optional,
-                        nullable: false,
-                    },
-                );
-            }
-            _ => {}
-        }
-    }
-}
-
-fn collect_props_from_mapped_type(
-    mapped: &oxc_ast::ast::TSMappedType<'_>,
-    source: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    seen: &mut Vec<String>,
-    props: &mut Vec<(String, PropTypeInfo)>,
-) {
-    let Some(keys) = literal_values_from_ts_type(
-        &mapped.constraint,
-        source,
-        interfaces,
-        type_aliases,
-        None,
-        seen,
-    ) else {
-        return;
-    };
-
-    let key_name = mapped.key.name.as_str();
-    for key in keys {
-        let prop_name = if let Some(name_type) = &mapped.name_type {
-            let mut values = literal_values_from_ts_type(
-                name_type,
-                source,
-                interfaces,
-                type_aliases,
-                Some((key_name, key.as_str())),
-                seen,
-            )
-            .unwrap_or_default();
-            if values.is_empty() {
-                continue;
-            }
-            values.remove(0)
-        } else {
-            key
-        };
-
-        let (ts_type, js_type, nullable) = if let Some(type_ann) = &mapped.type_annotation {
-            let ts_type = source_for_span(source, type_ann.span())
-                .unwrap_or("unknown")
-                .trim()
-                .to_compact_string();
-            let js_type = ts_type_to_js_type_from_ast(type_ann, source, interfaces, type_aliases);
-            let nullable = type_includes_top_level_null_from_ast(type_ann);
-            (Some(ts_type), js_type, nullable)
-        } else {
-            (None, "null".to_compact_string(), false)
-        };
-
-        push_prop_type_info(
-            props,
-            prop_name,
-            PropTypeInfo {
-                js_type,
-                ts_type,
-                optional: matches!(
-                    mapped.optional,
-                    Some(TSMappedTypeModifierOperator::True | TSMappedTypeModifierOperator::Plus)
-                ),
-                nullable,
-            },
-        );
-    }
-}
-
-fn push_prop_type_info(props: &mut Vec<(String, PropTypeInfo)>, name: String, info: PropTypeInfo) {
-    if !props
-        .iter()
-        .any(|(existing, _)| existing.as_str() == name.as_str())
-    {
-        props.push((name, info));
-    }
-}
-
-fn ts_type_to_js_type_from_ast(
-    ts_type: &TSType<'_>,
-    source: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-) -> String {
-    let mut seen = Vec::new();
-    ts_type_to_js_type_from_ast_inner(ts_type, source, interfaces, type_aliases, &mut seen)
-}
-
-/// Inner implementation of [`ts_type_to_js_type_from_ast`] that threads a
-/// `seen` stack of type-alias names currently being resolved. Self-referential
-/// or mutually recursive type aliases (e.g. `type NestedItem<T> = T extends
-/// Array<infer I> ? NestedItem<I> : T`, or a recursive `DotPathKeys`) otherwise
-/// re-parse and re-enter the same alias body forever, overflowing the native
-/// stack and aborting the whole `vize build`. The guard stops following an
-/// alias reference once its name is already on the stack and falls back to a
-/// safe `null` runtime type, mirroring the cycle protection already used by
-/// `collect_props_from_ts_type` / `resolve_type_reference_text`.
-fn ts_type_to_js_type_from_ast_inner(
-    ts_type: &TSType<'_>,
-    source: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    seen: &mut Vec<String>,
-) -> String {
-    match ts_type {
-        TSType::TSStringKeyword(_) => "String".to_compact_string(),
-        TSType::TSNumberKeyword(_) => "Number".to_compact_string(),
-        TSType::TSBooleanKeyword(_) => "Boolean".to_compact_string(),
-        TSType::TSSymbolKeyword(_) => "Symbol".to_compact_string(),
-        TSType::TSBigIntKeyword(_) => "BigInt".to_compact_string(),
-        TSType::TSObjectKeyword(_) | TSType::TSTypeLiteral(_) | TSType::TSMappedType(_) => {
-            "Object".to_compact_string()
-        }
-        TSType::TSArrayType(_) | TSType::TSTupleType(_) => "Array".to_compact_string(),
-        TSType::TSFunctionType(_) | TSType::TSConstructorType(_) => "Function".to_compact_string(),
-        TSType::TSLiteralType(lit) => js_type_for_ts_literal(&lit.literal),
-        TSType::TSUnionType(union) => {
-            combine_runtime_js_types(union.types.iter().filter_map(|ty| {
-                if matches!(
-                    ty,
-                    TSType::TSUndefinedKeyword(_)
-                        | TSType::TSNullKeyword(_)
-                        | TSType::TSNeverKeyword(_)
-                ) {
-                    None
-                } else {
-                    Some(ts_type_to_js_type_from_ast_inner(
-                        ty,
-                        source,
-                        interfaces,
-                        type_aliases,
-                        seen,
-                    ))
-                }
-            }))
-        }
-        TSType::TSConditionalType(conditional) => combine_runtime_js_types([
-            ts_type_to_js_type_from_ast_inner(
-                &conditional.true_type,
-                source,
-                interfaces,
-                type_aliases,
-                seen,
-            ),
-            ts_type_to_js_type_from_ast_inner(
-                &conditional.false_type,
-                source,
-                interfaces,
-                type_aliases,
-                seen,
-            ),
-        ]),
-        TSType::TSIntersectionType(intersection) => {
-            if intersection.types.iter().any(|ty| {
-                matches!(
-                    ty,
-                    TSType::TSTypeLiteral(_) | TSType::TSMappedType(_) | TSType::TSTypeReference(_)
-                )
-            }) {
-                "Object".to_compact_string()
-            } else {
-                "null".to_compact_string()
-            }
-        }
-        TSType::TSParenthesizedType(paren) => ts_type_to_js_type_from_ast_inner(
-            &paren.type_annotation,
-            source,
-            interfaces,
-            type_aliases,
-            seen,
-        ),
-        TSType::TSTypeOperatorType(op) => match op.operator {
-            TSTypeOperatorOperator::Readonly => ts_type_to_js_type_from_ast_inner(
-                &op.type_annotation,
-                source,
-                interfaces,
-                type_aliases,
-                seen,
-            ),
-            TSTypeOperatorOperator::Keyof => "String".to_compact_string(),
-            TSTypeOperatorOperator::Unique => "null".to_compact_string(),
-        },
-        TSType::TSTypeReference(type_ref) => {
-            let source_type = source_for_span(source, type_ref.span()).unwrap_or_default();
-            let js_type = ts_type_to_js_type(source_type);
-            if js_type != "null" {
-                return js_type;
-            }
-
-            let Some(name) = simple_type_name(&type_ref.type_name) else {
-                return "null".to_compact_string();
-            };
-            if interfaces.is_some_and(|interfaces| interfaces.contains_key(name)) {
-                return "Object".to_compact_string();
-            }
-            if let Some(type_aliases) = type_aliases
-                && let Some(alias) = type_aliases.get(name)
-            {
-                // Stop following recursive / mutually recursive aliases. Without
-                // this guard a self-referential alias body re-parses and
-                // re-enters here forever, overflowing the stack. (nuxt-ui
-                // `CheckboxGroup`/`RadioGroup` hit this via `NestedItem` /
-                // `DotPathKeys`.)
-                if seen.iter().any(|seen_name| seen_name == name) {
-                    return "null".to_compact_string();
-                }
-                let alias_source = wrap_type_alias_source(alias);
-                let allocator = Allocator::default();
-                let parsed = Parser::new(&allocator, &alias_source, SourceType::ts()).parse();
-                if !parsed.panicked
-                    && parsed.errors.is_empty()
-                    && let Some(Statement::TSTypeAliasDeclaration(alias_decl)) =
-                        parsed.program.body.first()
-                {
-                    seen.push(name.to_compact_string());
-                    let resolved = ts_type_to_js_type_from_ast_inner(
-                        &alias_decl.type_annotation,
-                        &alias_source,
-                        interfaces,
-                        Some(type_aliases),
-                        seen,
-                    );
-                    seen.pop();
-                    return resolved;
-                }
-            }
-            "null".to_compact_string()
-        }
-        TSType::TSIndexedAccessType(_) => "null".to_compact_string(),
-        _ => "null".to_compact_string(),
-    }
-}
-
-fn js_type_for_ts_literal(literal: &TSLiteral<'_>) -> String {
-    match literal {
-        TSLiteral::StringLiteral(_) | TSLiteral::TemplateLiteral(_) => "String".to_compact_string(),
-        TSLiteral::NumericLiteral(_) => "Number".to_compact_string(),
-        TSLiteral::BigIntLiteral(_) => "BigInt".to_compact_string(),
-        TSLiteral::BooleanLiteral(_) => "Boolean".to_compact_string(),
-        TSLiteral::UnaryExpression(unary) => {
-            if matches!(&unary.argument, Expression::NumericLiteral(_)) {
-                "Number".to_compact_string()
-            } else {
-                "null".to_compact_string()
-            }
-        }
-    }
-}
-
-fn combine_runtime_js_types(types: impl IntoIterator<Item = String>) -> String {
-    let mut js_types: Vec<String> = Vec::new();
-    for js_type in types {
-        if js_type == "null" {
-            return js_type;
-        }
-        if !js_types.contains(&js_type) {
-            js_types.push(js_type);
-        }
-    }
-
-    match js_types.len() {
-        0 => "null".to_compact_string(),
-        1 => js_types.pop().unwrap_or_else(|| "null".to_compact_string()),
-        _ => {
-            let joined = js_types.join(", ");
-            let mut result = String::with_capacity(joined.len() + 2);
-            result.push('[');
-            result.push_str(&joined);
-            result.push(']');
-            result
-        }
-    }
-}
-
-fn type_includes_top_level_null_from_ast(ts_type: &TSType<'_>) -> bool {
-    match ts_type {
-        TSType::TSNullKeyword(_) => true,
-        TSType::TSUnionType(union) => union
-            .types
-            .iter()
-            .any(|ty| matches!(ty, TSType::TSNullKeyword(_))),
-        TSType::TSParenthesizedType(paren) => {
-            type_includes_top_level_null_from_ast(&paren.type_annotation)
-        }
-        _ => false,
-    }
-}
-
-fn literal_values_from_ts_type(
-    ts_type: &TSType<'_>,
-    source: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    mapped_key: Option<(&str, &str)>,
-    seen: &mut Vec<String>,
-) -> Option<Vec<String>> {
-    match ts_type {
-        TSType::TSLiteralType(lit) => literal_values_from_literal(&lit.literal),
-        TSType::TSUnionType(union) => {
-            let mut values = Vec::new();
-            for ty in union.types.iter() {
-                values.extend(literal_values_from_ts_type(
-                    ty,
-                    source,
-                    interfaces,
-                    type_aliases,
-                    mapped_key,
-                    seen,
-                )?);
-            }
-            Some(values)
-        }
-        TSType::TSParenthesizedType(paren) => literal_values_from_ts_type(
-            &paren.type_annotation,
-            source,
-            interfaces,
-            type_aliases,
-            mapped_key,
-            seen,
-        ),
-        TSType::TSTemplateLiteralType(template) => literal_values_from_template_type(
-            template,
-            source,
-            interfaces,
-            type_aliases,
-            mapped_key,
-            seen,
-        ),
-        TSType::TSTypeOperatorType(op) if op.operator == TSTypeOperatorOperator::Keyof => {
-            literal_values_from_keyof_type(
-                &op.type_annotation,
-                source,
-                interfaces,
-                type_aliases,
-                seen,
-            )
-        }
-        TSType::TSTypeReference(type_ref) => {
-            let name = simple_type_name(&type_ref.type_name)?;
-            if let Some((mapped_name, mapped_value)) = mapped_key
-                && name == mapped_name
-            {
-                return Some(vec![mapped_value.to_compact_string()]);
-            }
-            let resolved = resolve_type_reference_text(name, interfaces, type_aliases, seen)?;
-            let values = literal_values_from_type_text(
-                &resolved,
-                interfaces,
-                type_aliases,
-                mapped_key,
-                seen,
-            );
-            finish_resolved_type_reference(name, seen);
-            values
-        }
-        _ => None,
-    }
-}
-
-fn literal_values_from_literal(literal: &TSLiteral<'_>) -> Option<Vec<String>> {
-    match literal {
-        TSLiteral::StringLiteral(lit) => Some(vec![lit.value.to_compact_string()]),
-        TSLiteral::NumericLiteral(lit) => Some(vec![lit.value.to_compact_string()]),
-        TSLiteral::TemplateLiteral(template) => {
-            literal_value_from_template_literal(template).map(|value| vec![value])
-        }
-        _ => None,
-    }
-}
-
-fn literal_values_from_template_type(
-    template: &oxc_ast::ast::TSTemplateLiteralType<'_>,
-    source: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    mapped_key: Option<(&str, &str)>,
-    seen: &mut Vec<String>,
-) -> Option<Vec<String>> {
-    let mut values = vec![String::default()];
-    for (idx, quasi) in template.quasis.iter().enumerate() {
-        let text = quasi
-            .value
-            .cooked
-            .as_ref()
-            .unwrap_or(&quasi.value.raw)
-            .as_str();
-        for value in &mut values {
-            value.push_str(text);
-        }
-        let Some(ty) = template.types.get(idx) else {
-            continue;
-        };
-        let parts =
-            literal_values_from_ts_type(ty, source, interfaces, type_aliases, mapped_key, seen)?;
-        let mut expanded = Vec::new();
-        for prefix in &values {
-            for part in &parts {
-                let mut value = prefix.clone();
-                value.push_str(part);
-                expanded.push(value);
-            }
-        }
-        values = expanded;
-    }
-    Some(values)
-}
-
-fn literal_values_from_keyof_type(
-    ts_type: &TSType<'_>,
-    source: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    seen: &mut Vec<String>,
-) -> Option<Vec<String>> {
-    let mut props = Vec::new();
-    if collect_props_from_ts_type(ts_type, source, interfaces, type_aliases, seen, &mut props) {
-        Some(props.into_iter().map(|(name, _)| name).collect())
-    } else {
-        None
-    }
-}
-
-fn literal_values_from_type_text(
-    type_text: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    mapped_key: Option<(&str, &str)>,
-    seen: &mut Vec<String>,
-) -> Option<Vec<String>> {
-    let source = wrap_type_alias_source(type_text);
-    let allocator = Allocator::default();
-    let parsed = Parser::new(&allocator, &source, SourceType::ts()).parse();
-    if parsed.panicked || !parsed.errors.is_empty() {
-        return None;
-    }
-    let Some(Statement::TSTypeAliasDeclaration(alias)) = parsed.program.body.first() else {
-        return None;
-    };
-    literal_values_from_ts_type(
-        &alias.type_annotation,
-        &source,
-        interfaces,
-        type_aliases,
-        mapped_key,
-        seen,
-    )
-}
-
-fn property_key_name(key: &PropertyKey<'_>) -> Option<String> {
-    match key {
-        PropertyKey::StaticIdentifier(id) => Some(id.name.to_compact_string()),
-        PropertyKey::StringLiteral(lit) => Some(lit.value.to_compact_string()),
-        PropertyKey::NumericLiteral(lit) => Some(lit.value.to_compact_string()),
-        PropertyKey::TemplateLiteral(template) => literal_value_from_template_literal(template),
-        _ => None,
-    }
-}
-
-fn literal_value_from_template_literal(
-    template: &oxc_ast::ast::TemplateLiteral<'_>,
-) -> Option<String> {
-    if !template.expressions.is_empty() {
-        return None;
-    }
-    let mut value = String::default();
-    for quasi in template.quasis.iter() {
-        value.push_str(
-            quasi
-                .value
-                .cooked
-                .as_ref()
-                .unwrap_or(&quasi.value.raw)
-                .as_str(),
-        );
-    }
-    Some(value)
-}
-
-fn source_for_span(source: &str, span: oxc_span::Span) -> Option<&str> {
-    source.get(span.start as usize..span.end as usize)
-}
-
-fn simple_type_name<'a>(type_name: &'a TSTypeName<'_>) -> Option<&'a str> {
-    match type_name {
-        TSTypeName::IdentifierReference(id) => Some(id.name.as_str()),
-        _ => None,
-    }
-}
-
-fn resolve_type_reference_text(
-    name: &str,
-    interfaces: Option<&FxHashMap<String, String>>,
-    type_aliases: Option<&FxHashMap<String, String>>,
-    seen: &mut Vec<String>,
-) -> Option<String> {
-    if seen.iter().any(|seen_name| seen_name == name) {
-        return None;
-    }
-
-    let resolved = if let Some(type_aliases) = type_aliases
-        && let Some(alias) = type_aliases.get(name)
-    {
-        Some(alias.clone())
-    } else if let Some(interfaces) = interfaces
-        && let Some(body) = interfaces.get(name)
-    {
-        let mut source = String::with_capacity(body.len() + 4);
-        source.push_str("{ ");
-        source.push_str(body);
-        source.push_str(" }");
-        Some(source)
-    } else {
-        None
-    }?;
-
-    seen.push(name.to_compact_string());
-    Some(resolved)
-}
-
-fn finish_resolved_type_reference(name: &str, seen: &mut Vec<String>) {
-    if seen.last().is_some_and(|seen_name| seen_name == name) {
-        seen.pop();
-    }
-}
-
-pub(crate) fn runtime_prop_key(name: &str) -> String {
-    if is_valid_identifier(name) {
-        return name.to_compact_string();
-    }
-
-    serde_json::to_string(name)
-        .map(|escaped| escaped.as_str().to_compact_string())
-        .unwrap_or_else(|_| {
-            let mut escaped = String::with_capacity(name.len() + 2);
-            escaped.push('"');
-            escaped.push_str(name);
-            escaped.push('"');
-            escaped
-        })
-}
-
-fn extract_prop_type_info(segment: &str, props: &mut Vec<(String, PropTypeInfo)>) {
-    let trimmed = segment.trim();
-    if trimmed.is_empty() {
-        return;
-    }
-
-    // Parse "name?: Type" or "name: Type"
-    if let Some(colon_pos) = trimmed.find(':') {
-        // Method-signature props (`onChange(e: E): void`, `update?(): T`)
-        // have a `(` somewhere in the parameter list *before* this colon.
-        // Detect by scanning the bytes up to `colon_pos` for `(`; on a hit,
-        // recover the name from before that `(` and treat the prop as
-        // `Function`-typed. Plain props (`name: Type`) skip this branch
-        // because there's no `(` before the colon. (#967)
-        let before_colon_bytes = &trimmed.as_bytes()[..colon_pos];
-        if let Some(paren_pos) = before_colon_bytes.iter().position(|&b| b == b'(') {
-            let before_paren = &trimmed[..paren_pos];
-            let optional = before_paren.trim_end().ends_with('?');
-            let name = before_paren.trim_end_matches('?').trim();
-            if !name.is_empty() && is_valid_identifier(name) {
-                let ts_type_str: String = (&trimmed[paren_pos..]).to_compact_string();
-                if !props.iter().any(|(n, _)| n == name) {
-                    props.push((
-                        name.to_compact_string(),
-                        PropTypeInfo {
-                            js_type: "Function".to_compact_string(),
-                            ts_type: Some(ts_type_str),
-                            optional,
-                            nullable: false,
-                        },
-                    ));
-                }
-            }
-            return;
-        }
-        let name_part = &trimmed[..colon_pos];
-        let type_part = &trimmed[colon_pos + 1..];
-
-        // Per Vue's type-only inference, a property is `required: false` only
-        // when the declaration carries the `?` optional modifier. `T |
-        // undefined` (no `?`) is still required. (#967)
-        let optional = name_part.ends_with('?');
-        let nullable = type_includes_top_level_null(type_part);
-        let name = name_part.trim().trim_end_matches('?').trim();
-
-        if !name.is_empty() && is_valid_identifier(name) {
-            let ts_type_str = type_part.trim().to_compact_string();
-            let js_type = ts_type_to_js_type(&ts_type_str);
-            // Avoid duplicates (intersection types may have overlapping props)
-            if !props.iter().any(|(n, _)| n == name) {
-                props.push((
-                    name.to_compact_string(),
-                    PropTypeInfo {
-                        js_type,
-                        ts_type: Some(ts_type_str),
-                        optional,
-                        nullable,
-                    },
-                ));
-            }
-        }
-    }
-}
-
-#[allow(dead_code)]
-fn type_includes_top_level_undefined(ts_type: &str) -> bool {
-    split_type_at_top_level(ts_type.trim(), '|')
-        .into_iter()
-        .any(|part| part.trim() == "undefined")
-}
-
-fn type_includes_top_level_null(ts_type: &str) -> bool {
-    split_type_at_top_level(ts_type.trim(), '|')
-        .into_iter()
-        .any(|part| part.trim() == "null")
-}
-
-pub fn add_null_to_runtime_type(js_type: &str, nullable: bool) -> String {
-    if !nullable || js_type == "null" {
-        return js_type.to_compact_string();
-    }
-
-    if js_type.starts_with('[') && js_type.ends_with(']') {
-        let inner = &js_type[1..js_type.len() - 1];
-        if inner
-            .split(',')
-            .map(|part| part.trim())
-            .any(|part| part == "null")
-        {
-            return js_type.to_compact_string();
-        }
-
-        let mut result = String::with_capacity(js_type.len() + 6);
-        result.push('[');
-        result.push_str(inner);
-        if !inner.trim().is_empty() {
-            result.push_str(", ");
-        }
-        result.push_str("null");
-        result.push(']');
-        return result;
-    }
-
-    let mut result = String::with_capacity(js_type.len() + 8);
-    result.push('[');
-    result.push_str(js_type);
-    result.push_str(", null]");
-    result
-}
-
-/// Split a type string at a delimiter only at the top level (depth 0),
-/// respecting nested `<>`, `()`, `[]`, `{}` and `=>` arrows.
-///
-/// This intentionally streams over `chars()` with a one-character lookbehind.
-/// Prop codegen calls it recursively for union members, so avoiding an
-/// intermediate `Vec<char>` prevents repeated heap churn in large prop types.
-fn split_type_at_top_level(s: &str, delimiter: char) -> Vec<String> {
-    let mut parts = Vec::new();
-    let mut current = String::default();
-    let mut depth: i32 = 0;
-    let mut prev = '\0';
-
-    for c in s.chars() {
-        match c {
-            '(' | '[' | '{' | '<' => {
-                depth += 1;
-                current.push(c);
-            }
-            ')' | ']' | '}' => {
-                if depth > 0 {
-                    depth -= 1;
-                }
-                current.push(c);
-            }
-            '>' => {
-                // Don't count > as closing angle bracket when preceded by = (arrow =>)
-                if prev == '=' {
-                    current.push(c);
-                } else {
-                    if depth > 0 {
-                        depth -= 1;
-                    }
-                    current.push(c);
-                }
-            }
-            c2 if c2 == delimiter && depth == 0 => {
-                parts.push(std::mem::take(&mut current));
-            }
-            _ => current.push(c),
-        }
-        prev = c;
-    }
-    if !current.is_empty() || !parts.is_empty() {
-        parts.push(current);
-    }
-    parts
-}
-
-/// Check if a type string contains a top-level `=>` (arrow function signature).
-///
-/// Like `split_type_at_top_level`, this is a zero-intermediate scanner because
-/// it sits on the recursive type-to-runtime-constructor path.
-fn contains_top_level_arrow(s: &str) -> bool {
-    let mut depth: i32 = 0;
-    let mut prev = '\0';
-    for c in s.chars() {
-        match c {
-            '(' | '[' | '{' | '<' => depth += 1,
-            ')' | ']' | '}' => {
-                if depth > 0 {
-                    depth -= 1;
-                }
-            }
-            '>' => {
-                if prev == '=' {
-                    // This is `=>`
-                    if depth == 0 {
-                        return true;
-                    }
-                    // Inside nested structure — don't change depth
-                } else if depth > 0 {
-                    depth -= 1;
-                }
-            }
-            _ => {}
-        }
-        prev = c;
-    }
-    false
-}
-
-/// Convert TypeScript type to JavaScript type constructor
-pub(crate) fn ts_type_to_js_type(ts_type: &str) -> String {
-    let ts_type = ts_type.trim();
-
-    // Strip `readonly` prefix: `readonly T[]` → `T[]`
-    let ts_type = if let Some(rest) = ts_type.strip_prefix("readonly ") {
-        rest.trim()
-    } else {
-        ts_type
-    };
-
-    // Handle string literal types: "foo" or 'bar' -> String
-    if (ts_type.starts_with('"') && ts_type.ends_with('"'))
-        || (ts_type.starts_with('\'') && ts_type.ends_with('\''))
-    {
-        return "String".to_compact_string();
-    }
-
-    // Handle numeric literal types: 123, 1.5 -> Number
-    if ts_type.parse::<f64>().is_ok() {
-        return "Number".to_compact_string();
-    }
-
-    // Handle boolean literal types: true, false -> Boolean
-    if ts_type == "true" || ts_type == "false" {
-        return "Boolean".to_compact_string();
-    }
-
-    // Arrow function types must be detected BEFORE union splitting,
-    // because `(x: T) => A | B` is a single function type (return type is `A | B`),
-    // not a union of `(x: T) => A` and `B`.
-    // Also must come before array/object checks because `(items: T[]) => T[]`
-    // ends with `[]` and contains `:`.
-    if contains_top_level_arrow(ts_type) {
-        return "Function".to_compact_string();
-    }
-
-    // Handle union types — split at top level only (respecting nesting).
-    // For mixed types like `string | number`, produce `[String, Number]`.
-    {
-        let parts = split_type_at_top_level(ts_type, '|');
-        if parts.len() > 1 {
-            let meaningful: Vec<&str> = parts
-                .iter()
-                .map(|p| p.trim())
-                .filter(|p| !p.is_empty() && *p != "undefined" && *p != "null")
-                .collect();
-
-            if meaningful.is_empty() {
-                return "null".to_compact_string();
-            }
-
-            // Collect unique JS types for each union member
-            let mut js_types: Vec<String> = Vec::new();
-            for part in &meaningful {
-                let jt = ts_type_to_js_type(part);
-                if !js_types.contains(&jt) {
-                    if jt == "null" {
-                        return jt;
-                    }
-                    js_types.push(jt);
-                }
-            }
-
-            if js_types.len() == 1
-                && let Some(only) = js_types.pop()
-            {
-                return only;
-            }
-
-            // Multiple distinct types → array form: [String, Number]
-            let joined = js_types.join(", ");
-            let mut result = String::with_capacity(joined.len() + 2);
-            result.push('[');
-            result.push_str(&joined);
-            result.push(']');
-            return result;
-        }
-    }
-
-    // Map TypeScript types to JavaScript constructors
-    match ts_type.to_lowercase().as_str() {
-        "string" => "String".to_compact_string(),
-        "number" => "Number".to_compact_string(),
-        "boolean" => "Boolean".to_compact_string(),
-        "object" => "Object".to_compact_string(),
-        "function" => "Function".to_compact_string(),
-        "symbol" => "Symbol".to_compact_string(),
-        _ => {
-            // Handle array types
-            if ts_type.ends_with("[]") || ts_type.starts_with("Array<") {
-                "Array".to_compact_string()
-            } else if ts_type.starts_with('{') || contains_top_level_colon(ts_type) {
-                // Object literal type
-                "Object".to_compact_string()
-            } else if ts_type.starts_with('(') && ts_type.contains("=>") {
-                // Function type (fallback, already handled above)
-                "Function".to_compact_string()
-            } else {
-                // Check if this is a built-in JavaScript constructor type
-                let type_name = ts_type.split('<').next().unwrap_or(ts_type).trim();
-                match type_name {
-                    // Built-in JavaScript types that exist at runtime
-                    "Date" | "RegExp" | "Error" | "Map" | "Set" | "WeakMap" | "WeakSet"
-                    | "Promise" | "ArrayBuffer" | "DataView" | "Int8Array" | "Uint8Array"
-                    | "Int16Array" | "Uint16Array" | "Int32Array" | "Uint32Array"
-                    | "Float32Array" | "Float64Array" | "BigInt64Array" | "BigUint64Array"
-                    | "URL" | "URLSearchParams" | "FormData" | "Blob" | "File" => {
-                        type_name.to_compact_string()
-                    }
-                    // Vue reactive types that are objects at runtime
-                    "Ref"
-                    | "ShallowRef"
-                    | "ComputedRef"
-                    | "WritableComputedRef"
-                    | "MaybeRef"
-                    | "MaybeRefOrGetter"
-                    | "Readonly"
-                    | "UnwrapRef"
-                    | "Reactive"
-                    | "ShallowReactive"
-                    | "ToRef"
-                    | "ToRefs" => "Object".to_compact_string(),
-                    // User-defined interface/type or generic type parameter
-                    // - Single uppercase letter (T, U, K, V) = generic param → null
-                    // - Otherwise = user-defined type → null (types don't exist at runtime)
-                    _ => "null".to_compact_string(),
-                }
-            }
-        }
-    }
-}
-
-/// Check if a type string contains a `:` at the top level (not inside generics/parens).
-/// Used to detect object literal types like `{ key: string }` vs types like `Record<K, V>`.
-fn contains_top_level_colon(s: &str) -> bool {
-    let mut depth: i32 = 0;
-    let mut prev = '\0';
-    for c in s.chars() {
-        match c {
-            '(' | '[' | '{' | '<' => depth += 1,
-            ')' | ']' | '}' => {
-                if depth > 0 {
-                    depth -= 1;
-                }
-            }
-            '>' => {
-                if prev == '=' {
-                    // Arrow =>, don't change depth
-                } else if depth > 0 {
-                    depth -= 1;
-                }
-            }
-            ':' if depth == 0 => return true,
-            _ => {}
-        }
-        prev = c;
-    }
-    false
-}
-
-/// Resolve prop type references using type alias/interface maps.
-/// For a prop type like `ButtonVariant`, resolves it using the type_aliases and interfaces
-/// to determine the correct JS type constructor.
-pub fn resolve_prop_js_type(
-    ts_type: &str,
-    interfaces: &FxHashMap<String, String>,
-    type_aliases: &FxHashMap<String, String>,
-) -> Option<String> {
-    let trimmed = ts_type.trim();
-    // Check if it's a simple type reference (identifier, no generics/brackets/arrows/pipes)
-    // that would resolve to `null` by default
-    if trimmed.is_empty() {
-        return None;
-    }
-
-    // First try the normal resolution
-    let js_type = ts_type_to_js_type(trimmed);
-    if js_type != "null" {
-        return None; // Normal resolution works fine
-    }
-
-    // It resolved to null - try to look up the type name and resolve based on the actual definition
-    let base_name = if let Some(idx) = trimmed.find('<') {
-        trimmed[..idx].trim()
-    } else {
-        trimmed
-    };
-
-    // Look up in type aliases first
-    if let Some(body) = type_aliases.get(base_name) {
-        let resolved_type = ts_type_to_js_type(body.trim());
-        if resolved_type != "null" {
-            return Some(resolved_type);
-        }
-        // If the alias body contains braces, it's an object type
-        if body.contains('{') {
-            return Some("Object".to_compact_string());
-        }
-    }
-
-    // Look up in interfaces
-    if let Some(body) = interfaces.get(base_name) {
-        // Interfaces always resolve to Object
-        let _ = body;
-        return Some("Object".to_compact_string());
-    }
-
-    None
-}
-
-/// Strip the `readonly` keyword from a TypeScript type.
-/// Handles patterns like `readonly { value: string }[]` → `{ value: string }[]`
-pub fn strip_readonly_prefix(ts_type: &str) -> &str {
-    let trimmed = ts_type.trim();
-    if let Some(rest) = trimmed.strip_prefix("readonly ") {
-        rest.trim()
-    } else {
-        trimmed
-    }
-}
-
-/// Extract emit names from TypeScript type definition
-pub fn extract_emit_names_from_type(type_args: &str) -> Vec<String> {
-    let mut emits = Vec::new();
-
-    // First, try Vue 3.3+ shorthand format:
-    //   { change: [value: string]; submit: []; update: [id: number] }
-    // Property names before `:` followed by `[` are event names
-    let trimmed = type_args.trim();
-    let is_shorthand = trimmed.starts_with('{')
-        && trimmed.contains('[')
-        && !trimmed.contains("(e:")
-        && !trimmed.contains("(event:");
-
-    if is_shorthand {
-        // Extract property names from { name: [...], name: [...] } format
-        let inner = if trimmed.starts_with('{') && trimmed.ends_with('}') {
-            &trimmed[1..trimmed.len() - 1]
-        } else {
-            trimmed
-        };
-
-        // Split by lines or semicolons and extract property names.
-        for segment in inner.split([';', '\n']) {
-            if let Some(name) = extract_emit_shorthand_key(segment) {
-                emits.push(name);
-            }
-        }
-
-        if !emits.is_empty() {
-            return emits;
-        }
-    }
-
-    extract_call_signature_emit_names(type_args, &mut emits);
-    emits
-}
-
-fn extract_call_signature_emit_names(type_args: &str, emits: &mut Vec<String>) {
-    let bytes = type_args.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        if bytes[i] != b'(' {
-            i += 1;
-            continue;
-        }
-
-        let Some(close) = find_matching_paren(type_args, i) else {
-            i += 1;
-            continue;
-        };
-
-        if is_emit_call_signature(type_args, i, close) {
-            let params = &type_args[i + 1..close];
-            if let Some(first_param) = first_parameter(params)
-                && let Some(type_annotation) = parameter_type_annotation(first_param)
-            {
-                extract_string_literals(type_annotation, emits);
-            }
-        }
-
-        i = close + 1;
-    }
-}
-
-fn find_matching_paren(input: &str, open: usize) -> Option<usize> {
-    let bytes = input.as_bytes();
-    let mut depth = 0;
-    let mut i = open;
-    let mut quote: Option<u8> = None;
-
-    while i < bytes.len() {
-        let byte = bytes[i];
-        if let Some(quote_byte) = quote {
-            if byte == b'\\' {
-                i += 2;
-                continue;
-            }
-            if byte == quote_byte {
-                quote = None;
-            }
-            i += 1;
-            continue;
-        }
-
-        match byte {
-            b'\'' | b'"' => quote = Some(byte),
-            b'(' => depth += 1,
-            b')' => {
-                depth -= 1;
-                if depth == 0 {
-                    return Some(i);
-                }
-            }
-            _ => {}
-        }
-        i += 1;
-    }
-
-    None
-}
-
-fn is_emit_call_signature(input: &str, open: usize, close: usize) -> bool {
-    let after = input[close + 1..].trim_start();
-    if after.starts_with(':') {
-        return true;
-    }
-
-    // Direct function type form: defineEmits<(e: 'change') => void>().
-    after.starts_with("=>") && input[..open].trim().is_empty()
-}
-
-fn first_parameter(params: &str) -> Option<&str> {
-    let mut depth = 0;
-    let mut quote: Option<char> = None;
-    let mut prev_escape = false;
-
-    for (idx, ch) in params.char_indices() {
-        if let Some(quote_char) = quote {
-            if prev_escape {
-                prev_escape = false;
-                continue;
-            }
-            if ch == '\\' {
-                prev_escape = true;
-                continue;
-            }
-            if ch == quote_char {
-                quote = None;
-            }
-            continue;
-        }
-
-        match ch {
-            '\'' | '"' => quote = Some(ch),
-            '(' | '[' | '{' | '<' => depth += 1,
-            ')' | ']' | '}' | '>' if depth > 0 => depth -= 1,
-            ',' if depth == 0 => return Some(params[..idx].trim()),
-            _ => {}
-        }
-    }
-
-    let first = params.trim();
-    (!first.is_empty()).then_some(first)
-}
-
-fn parameter_type_annotation(param: &str) -> Option<&str> {
-    let mut depth = 0;
-    let mut quote: Option<char> = None;
-    let mut prev_escape = false;
-
-    for (idx, ch) in param.char_indices() {
-        if let Some(quote_char) = quote {
-            if prev_escape {
-                prev_escape = false;
-                continue;
-            }
-            if ch == '\\' {
-                prev_escape = true;
-                continue;
-            }
-            if ch == quote_char {
-                quote = None;
-            }
-            continue;
-        }
-
-        match ch {
-            '\'' | '"' => quote = Some(ch),
-            '(' | '[' | '{' | '<' => depth += 1,
-            ')' | ']' | '}' | '>' if depth > 0 => depth -= 1,
-            ':' if depth == 0 => return Some(param[idx + 1..].trim()),
-            _ => {}
-        }
-    }
-
-    None
-}
-
-fn extract_string_literals(input: &str, output: &mut Vec<String>) {
-    let bytes = input.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        let quote = bytes[i];
-        if !matches!(quote, b'\'' | b'"') {
-            i += 1;
-            continue;
-        }
-
-        let mut literal = String::default();
-        i += 1;
-        while i < bytes.len() {
-            let byte = bytes[i];
-            if byte == b'\\' {
-                if i + 1 < bytes.len() {
-                    literal.push(bytes[i + 1] as char);
-                    i += 2;
-                    continue;
-                }
-                break;
-            }
-            if byte == quote {
-                if !literal.is_empty() {
-                    output.push(literal);
-                }
-                i += 1;
-                break;
-            }
-            literal.push(byte as char);
-            i += 1;
-        }
-    }
-}
-
-fn extract_emit_shorthand_key(segment: &str) -> Option<String> {
-    let seg = segment.trim();
-    if seg.is_empty() || seg.starts_with("...") {
-        return None;
-    }
-
-    let bytes = seg.as_bytes();
-    let first = *bytes.first()?;
-    if matches!(first, b'\'' | b'"' | b'`') {
-        let quote = first;
-        let mut key = String::default();
-        let mut i = 1;
-        while i < bytes.len() {
-            let c = bytes[i];
-            if c == b'\\' {
-                if i + 1 < bytes.len() {
-                    key.push(bytes[i + 1] as char);
-                    i += 2;
-                    continue;
-                }
-                return None;
-            }
-            if c == quote {
-                let rest = seg[i + 1..].trim_start();
-                if rest.starts_with(':') && !key.is_empty() {
-                    return Some(key);
-                }
-                return None;
-            }
-            key.push(c as char);
-            i += 1;
-        }
-        return None;
-    }
-
-    let mut colon_pos = None;
-    for (idx, ch) in seg.char_indices() {
-        if ch == ':' {
-            colon_pos = Some(idx);
-            break;
-        }
-    }
-    let name = seg[..colon_pos?].trim().trim_end_matches('?').trim();
-    if !name.is_empty()
-        && name
-            .chars()
-            .all(|c| c.is_alphanumeric() || c == '_' || c == '-' || c == '$')
-    {
-        Some(name.to_compact_string())
-    } else {
-        None
-    }
-}
-
-/// Extract default values from withDefaults second argument
-/// Input: "withDefaults(defineProps<{...}>(), { prop1: default1, prop2: default2 })"
-/// Returns: HashMap of prop name to default value string
-pub fn extract_with_defaults_defaults(with_defaults_args: &str) -> FxHashMap<String, String> {
-    let mut defaults = FxHashMap::default();
-    let trimmed = with_defaults_args.trim();
-    if trimmed.is_empty() {
-        return defaults;
-    }
-
-    const WRAP_PREFIX: &str = "const __vize_defaults__ = ";
-    let mut wrapped = String::with_capacity(WRAP_PREFIX.len() + trimmed.len() + 1);
-    wrapped.push_str(WRAP_PREFIX);
-    wrapped.push_str(trimmed);
-    wrapped.push(';');
-
-    let allocator = Allocator::default();
-    let parse_result = Parser::new(
-        &allocator,
-        &wrapped,
-        SourceType::default().with_typescript(true),
-    )
-    .parse();
-    if !parse_result.errors.is_empty() {
-        return defaults;
-    }
-
-    let Some(Statement::VariableDeclaration(var_decl)) = parse_result.program.body.first() else {
-        return defaults;
-    };
-    let Some(declarator) = var_decl.declarations.first() else {
-        return defaults;
-    };
-    let Some(Expression::CallExpression(call)) = declarator.init.as_ref() else {
-        return defaults;
-    };
-    let Expression::Identifier(callee) = &call.callee else {
-        return defaults;
-    };
-    if callee.name.as_str() != "withDefaults" {
-        return defaults;
-    }
-
-    let Some(Argument::ObjectExpression(obj)) = call.arguments.get(1) else {
-        return defaults;
-    };
-
-    for property in obj.properties.iter() {
-        let ObjectPropertyKind::ObjectProperty(prop) = property else {
-            continue;
-        };
-
-        let key = match &prop.key {
-            PropertyKey::StaticIdentifier(id) => id.name.to_compact_string(),
-            PropertyKey::StringLiteral(lit) => lit.value.to_compact_string(),
-            PropertyKey::NumericLiteral(lit) => lit.value.to_compact_string(),
-            _ => continue,
-        };
-
-        let Some(value_start) = (prop.value.span().start as usize).checked_sub(WRAP_PREFIX.len())
-        else {
-            continue;
-        };
-        let Some(value_end) = (prop.value.span().end as usize).checked_sub(WRAP_PREFIX.len())
-        else {
-            continue;
-        };
-        if let Some(value_src) = trimmed.get(value_start..value_end) {
-            defaults.insert(key, value_src.to_compact_string());
-        }
-    }
-
-    defaults
-}
-
-/// Normalize default values from reactive props destructure for runtime props.
-///
-/// Vue treats array/object destructure defaults as per-instance factories, while
-/// function defaults are already factories/values and must not be wrapped.
-pub(crate) fn normalize_destructure_default_value(default_value: &str) -> String {
-    let trimmed = default_value.trim();
-    if trimmed.starts_with('[') {
-        let mut wrapped = String::with_capacity(trimmed.len() + 6);
-        wrapped.push_str("() => ");
-        wrapped.push_str(trimmed);
-        return wrapped;
-    }
-
-    if trimmed.starts_with('{') {
-        let mut wrapped = String::with_capacity(trimmed.len() + 8);
-        wrapped.push_str("() => (");
-        wrapped.push_str(trimmed);
-        wrapped.push(')');
-        return wrapped;
-    }
-
-    default_value.to_compact_string()
-}
-
-/// Validate Vue-specific script-setup semantics that the TypeScript checker
-/// cannot derive on its own (currently: prop destructure defaults whose value
-/// disagrees with the declared prop type).
-///
-/// This runs only the analysis needed to drive the validators — no codegen,
-/// no template compile — so check/LSP can call it on every SFC without paying
-/// the cost of `compile_sfc`. Returns `Ok(())` when there is no `<script
-/// setup>` or no actionable issue.
-pub fn validate_script_setup_semantics(script_setup_content: &str) -> Result<(), SfcError> {
-    // Delegating with a zero block offset yields block-relative diagnostic
-    // locations. Callers that know where the `<script setup>` block sits in
-    // the full SFC should prefer `validate_script_setup_semantics_located` so
-    // diagnostics map to the correct document position.
-    validate_script_setup_semantics_located(script_setup_content, 0, script_setup_content)
-}
-
-/// Like [`validate_script_setup_semantics`], but rebases diagnostic locations
-/// onto the full SFC: `block_start` is the byte offset of the script-setup
-/// block content within `sfc_source` (i.e. `script_setup.loc.start`), and
-/// `sfc_source` is the whole SFC text used to resolve line/column. This lets
-/// editor/check diagnostics point at the offending span (e.g. a destructure
-/// default) instead of the start of the `<script setup>` block.
-pub fn validate_script_setup_semantics_located(
-    script_setup_content: &str,
-    block_start: usize,
-    sfc_source: &str,
-) -> Result<(), SfcError> {
-    if script_setup_content.is_empty() {
-        return Ok(());
-    }
-    let mut ctx = ScriptCompileContext::new(script_setup_content);
-    ctx.analyze();
-    validate_props_destructure_default_types(&ctx, block_start, sfc_source)
-}
-
-/// Build a 1-based [`BlockLocation`] for an absolute byte span in `source`.
-/// Columns count UTF-16 code units to match the LSP position convention.
-fn block_location_for_span(source: &str, start: usize, end: usize) -> crate::types::BlockLocation {
-    fn line_col_1based(source: &str, offset: usize) -> (usize, usize) {
-        let target = offset.min(source.len());
-        let mut line = 1usize;
-        let mut column = 1usize;
-        for (index, ch) in source.char_indices() {
-            if index >= target {
-                break;
-            }
-            if ch == '\n' {
-                line += 1;
-                column = 1;
-            } else {
-                column += ch.len_utf16();
-            }
-        }
-        (line, column)
-    }
-    let (start_line, start_column) = line_col_1based(source, start);
-    let (end_line, end_column) = line_col_1based(source, end);
-    crate::types::BlockLocation {
-        start,
-        end,
-        tag_start: start,
-        tag_end: end,
-        start_line,
-        start_column,
-        end_line,
-        end_column,
-    }
-}
-
-/// Locate the byte span `(start, end)` of a reactive props-destructure
-/// default's value expression for `prop_key`, relative to
-/// `script_setup_content` (e.g. the `0` in `const { msg = 0 } = defineProps<…>()`).
-///
-/// Used to point the default-type-mismatch diagnostic at the offending
-/// default rather than the start of the `<script setup>` block. Re-parses the
-/// script-setup content; only the rare diagnostic error path calls this, so
-/// the extra parse is not on the hot path. The span lives on the AST, not on
-/// the public `PropsDestructureBinding`, to keep that type's API stable.
-fn props_destructure_default_span(
-    script_setup_content: &str,
-    prop_key: &str,
-) -> Option<(u32, u32)> {
-    use oxc_ast::ast::BindingPattern;
-
-    let allocator = Allocator::default();
-    let parsed = Parser::new(&allocator, script_setup_content, SourceType::ts()).parse();
-    for stmt in &parsed.program.body {
-        let Statement::VariableDeclaration(var_decl) = stmt else {
-            continue;
-        };
-        for decl in &var_decl.declarations {
-            // Match `const { … } = defineProps(…)`.
-            let BindingPattern::ObjectPattern(obj_pat) = &decl.id else {
-                continue;
-            };
-            let is_define_props = matches!(
-                &decl.init,
-                Some(Expression::CallExpression(call))
-                    if matches!(&call.callee, Expression::Identifier(id) if id.name == "defineProps")
-            );
-            if !is_define_props {
-                continue;
-            }
-            for prop in obj_pat.properties.iter() {
-                let key = match &prop.key {
-                    PropertyKey::StaticIdentifier(id) => id.name.as_str(),
-                    PropertyKey::StringLiteral(lit) => lit.value.as_str(),
-                    _ => continue,
-                };
-                if key != prop_key {
-                    continue;
-                }
-                if let BindingPattern::AssignmentPattern(assign) = &prop.value {
-                    let span = assign.right.span();
-                    return Some((span.start, span.end));
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Validate reactive props destructure defaults against inferred runtime prop types.
-pub(crate) fn validate_props_destructure_default_types(
-    ctx: &ScriptCompileContext,
-    block_start: usize,
-    sfc_source: &str,
-) -> Result<(), SfcError> {
-    let Some(destructure) = ctx.macros.props_destructure.as_ref() else {
-        return Ok(());
-    };
-    let Some(props_macro) = ctx.macros.define_props.as_ref() else {
-        return Ok(());
-    };
-    let Some(type_args) = props_macro.type_args.as_ref() else {
-        return Ok(());
-    };
-
-    let resolved_type_args =
-        crate::script::resolve_type_args(type_args, &ctx.interfaces, &ctx.type_aliases);
-    let prop_types = extract_prop_types_from_type_with_context(
-        &resolved_type_args,
-        Some(&ctx.interfaces),
-        Some(&ctx.type_aliases),
-    );
-
-    for (name, prop_type) in &prop_types {
-        let Some(binding) = destructure.bindings.get(name.as_str()) else {
-            continue;
-        };
-        let Some(default_value) = binding.default.as_deref() else {
-            continue;
-        };
-
-        let resolved_js_type = if prop_type.js_type == "null" {
-            prop_type
-                .ts_type
-                .as_ref()
-                .and_then(|ts_type| {
-                    resolve_prop_js_type(ts_type, &ctx.interfaces, &ctx.type_aliases)
-                })
-                .unwrap_or_else(|| prop_type.js_type.clone())
-        } else {
-            prop_type.js_type.clone()
-        };
-
-        if resolved_js_type == "null" || prop_type.nullable {
-            continue;
-        }
-
-        let Some(default_type) =
-            infer_default_value_runtime_type(default_value, resolved_js_type.as_str())
-        else {
-            continue;
-        };
-
-        if !runtime_type_includes(resolved_js_type.as_str(), default_type) {
-            // Point the diagnostic at the offending default expression when we
-            // can locate it, rebasing the block-relative span onto the full
-            // SFC. Falls back to no location (callers then pick a block-start
-            // fallback) only when the span cannot be resolved.
-            let loc =
-                props_destructure_default_span(&ctx.source, name).map(|(rel_start, rel_end)| {
-                    block_location_for_span(
-                        sfc_source,
-                        block_start + rel_start as usize,
-                        block_start + rel_end as usize,
-                    )
-                });
-            return Err(SfcError {
-                message: {
-                    let mut message = String::from("Default value of prop \"");
-                    message.push_str(name);
-                    message.push_str("\" does not match declared type.");
-                    message
-                },
-                code: Some("DEFINE_PROPS_DESTRUCTURE_DEFAULT_TYPE".into()),
-                loc,
-            });
-        }
-    }
-
-    Ok(())
-}
-
-fn infer_default_value_runtime_type(
-    default_value: &str,
-    expected_runtime_type: &str,
-) -> Option<&'static str> {
-    const WRAP_PREFIX: &str = "const __vize_default__ = ";
-    let mut wrapped = String::with_capacity(WRAP_PREFIX.len() + default_value.len() + 1);
-    wrapped.push_str(WRAP_PREFIX);
-    wrapped.push_str(default_value);
-    wrapped.push(';');
-
-    let allocator = Allocator::default();
-    let parse_result = Parser::new(
-        &allocator,
-        &wrapped,
-        SourceType::default().with_typescript(true),
-    )
-    .parse();
-    if !parse_result.errors.is_empty() {
-        return None;
-    }
-
-    let Some(Statement::VariableDeclaration(var_decl)) = parse_result.program.body.first() else {
-        return None;
-    };
-    let declarator = var_decl.declarations.first()?;
-    let value = declarator.init.as_ref()?;
-
-    infer_expression_runtime_type(unwrap_ts_expression(value), expected_runtime_type)
-}
-
-fn unwrap_ts_expression<'a>(expr: &'a Expression<'a>) -> &'a Expression<'a> {
-    match expr {
-        Expression::ParenthesizedExpression(paren) => unwrap_ts_expression(&paren.expression),
-        Expression::TSAsExpression(ts_as) => unwrap_ts_expression(&ts_as.expression),
-        Expression::TSSatisfiesExpression(ts_satisfies) => {
-            unwrap_ts_expression(&ts_satisfies.expression)
-        }
-        Expression::TSNonNullExpression(ts_non_null) => {
-            unwrap_ts_expression(&ts_non_null.expression)
-        }
-        _ => expr,
-    }
-}
-
-fn infer_expression_runtime_type(
-    expr: &Expression<'_>,
-    expected_runtime_type: &str,
-) -> Option<&'static str> {
-    match expr {
-        Expression::StringLiteral(_) => Some("String"),
-        Expression::TemplateLiteral(template) if template.expressions.is_empty() => Some("String"),
-        Expression::NumericLiteral(_) => Some("Number"),
-        Expression::BooleanLiteral(_) => Some("Boolean"),
-        Expression::ObjectExpression(_) => Some("Object"),
-        Expression::ArrayExpression(_) => Some("Array"),
-        Expression::ArrowFunctionExpression(arrow) => {
-            if runtime_type_includes(expected_runtime_type, "Function") {
-                return Some("Function");
-            }
-            arrow_return_expression(arrow)
-                .and_then(|expr| infer_expression_runtime_type(expr, expected_runtime_type))
-        }
-        Expression::FunctionExpression(func) => {
-            if runtime_type_includes(expected_runtime_type, "Function") {
-                return Some("Function");
-            }
-            function_body_return_expression(&func.body.as_ref()?.statements)
-                .and_then(|expr| infer_expression_runtime_type(expr, expected_runtime_type))
-        }
-        _ => None,
-    }
-}
-
-fn arrow_return_expression<'a>(
-    arrow: &'a oxc_ast::ast::ArrowFunctionExpression<'a>,
-) -> Option<&'a Expression<'a>> {
-    if !arrow.expression {
-        return function_body_return_expression(&arrow.body.statements);
-    }
-
-    let Statement::ExpressionStatement(expr_stmt) = arrow.body.statements.first()? else {
-        return None;
-    };
-    Some(&expr_stmt.expression)
-}
-
-fn function_body_return_expression<'a>(
-    statements: &'a oxc_allocator::Vec<'a, Statement<'a>>,
-) -> Option<&'a Expression<'a>> {
-    for stmt in statements.iter() {
-        if let Statement::ReturnStatement(ret) = stmt
-            && let Some(argument) = &ret.argument
-        {
-            return Some(argument);
-        }
-    }
-    None
-}
-
-fn runtime_type_includes(runtime_type: &str, value_type: &str) -> bool {
-    runtime_type
-        .trim_matches(|c| c == '[' || c == ']')
-        .split(',')
-        .map(str::trim)
-        .any(|part| part == value_type)
-}
-
-/// Check if a string is a valid JS identifier
-pub fn is_valid_identifier(s: &str) -> bool {
-    if s.is_empty() {
-        return false;
-    }
-
-    let mut chars = s.chars();
-    match chars.next() {
-        Some(c) if c.is_alphabetic() || c == '_' || c == '$' => {}
-        _ => return false,
-    }
-
-    chars.all(|c| c.is_alphanumeric() || c == '_' || c == '$')
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        extract_emit_names_from_type, extract_prop_types_from_type_with_context, ts_type_to_js_type,
-    };
-    use vize_carton::{FxHashMap, ToCompactString};
-
-    #[test]
-    fn recursive_type_alias_prop_does_not_overflow() {
-        // Mirrors nuxt-ui `CheckboxGroup`/`RadioGroup`: a props type whose member
-        // resolves through a self-referential alias (`DotPathKeys`) and a
-        // conditional/mutually-recursive alias (`NestedItem`). Before the cycle
-        // guard in `ts_type_to_js_type_from_ast`, resolving the runtime type of
-        // such a prop re-parsed the same alias body forever and overflowed the
-        // native stack, aborting `vize build`.
-        let mut type_aliases: FxHashMap<vize_carton::String, vize_carton::String> =
-            FxHashMap::default();
-        type_aliases.insert(
-            "NestedItem".to_compact_string(),
-            "T extends Array<infer I> ? NestedItem<I> : T".to_compact_string(),
-        );
-        type_aliases.insert(
-            "DotPathKeys".to_compact_string(),
-            "{ [K in keyof T & string]: K | `${K}.${DotPathKeys<T[K]>}` }[keyof T & string]"
-                .to_compact_string(),
-        );
-
-        let type_args = "{ items?: NestedItem<string>; keys?: DotPathKeys<Record<string, any>> }";
-
-        let props = extract_prop_types_from_type_with_context(type_args, None, Some(&type_aliases));
-
-        let items = props
-            .iter()
-            .find(|(name, _)| name == "items")
-            .expect("items prop should be extracted");
-        // Recursive structural types have no runtime constructor.
-        assert_eq!(items.1.js_type.as_str(), "null");
-        assert!(items.1.optional);
-
-        let keys = props
-            .iter()
-            .find(|(name, _)| name == "keys")
-            .expect("keys prop should be extracted");
-        assert_eq!(keys.1.js_type.as_str(), "null");
-    }
-
-    #[test]
-    fn extract_emit_names_keeps_quoted_colon_keys() {
-        let names = extract_emit_names_from_type(
-            r#"{
-              "update:open": [value: boolean]
-              'select:item': [id: string]
-              close: []
-            }"#,
-        );
-
-        assert_eq!(names, vec!["update:open", "select:item", "close"]);
-    }
-
-    #[test]
-    fn extract_emit_names_ignores_payload_literal_union_in_call_signature() {
-        let names = extract_emit_names_from_type("{ (e: 'change', mode: 'x' | 'y'): void }");
-
-        assert_eq!(names, vec!["change"]);
-    }
-
-    #[test]
-    fn extract_emit_names_ignores_payload_literal_union_in_function_type() {
-        let names = extract_emit_names_from_type("(e: 'change', mode: 'x' | 'y') => void");
-
-        assert_eq!(names, vec!["change"]);
-    }
-
-    #[test]
-    fn extract_emit_names_supports_custom_first_parameter_name() {
-        let names = extract_emit_names_from_type("{ (evt: 'change', val: 'a' | 'b'): void }");
-
-        assert_eq!(names, vec!["change"]);
-    }
-
-    #[test]
-    fn extract_emit_names_ignores_payload_literals_across_multiple_call_signatures() {
-        let names = extract_emit_names_from_type(
-            r#"{
-              (e: 'change', mode: 'x' | 'y'): void
-              (evt: 'submit', val: 'a' | 'b'): void
-            }"#,
-        );
-
-        assert_eq!(names, vec!["change", "submit"]);
-    }
-
-    #[test]
-    fn generic_union_with_array_falls_back_to_unknown_runtime_type() {
-        assert_eq!(ts_type_to_js_type("T | T[]"), "null");
-    }
-
-    #[test]
-    fn leading_union_separator_does_not_add_null_runtime_type() {
-        let ty = r#"
-          | { type: "link"; href: string }
-          | { type: "button"; onClick: () => void }
-        "#;
-
-        assert_eq!(ts_type_to_js_type(ty), "Object");
-    }
-}
